Create denyAllNotAdmin
This policy uses the "Deny" effect to block all access to all resources by all principals (users and roles) that do not have the "admin" group.
You will need to replace the "IAM_USER_ARN_1", "IAM_USER_ARN_2", "IAM_USER_ARN_3" with the actual arn of the users you want to be able to connect.

This is a basic example, you may need to customize the policy to fit your specific use case. For example, you may want to allow other actions or resources to be accessible by non-admin users.

Also, you may need to attach this policy to a specific resource or service like EC2, S3, RDS, etc.